{{cookiecutter.project_name}}/requirements/local.txt
-r base.txt

# main packages
psycopg2==2.9.6

# development tools
pre-commit==3.3.2
ruff==0.0.272
mypy==1.3.0
django-stubs==4.2.1
djangorestframework-stubs==3.14.1

django-debug-toolbar==4.1.0
django-extensions==3.2.3
ipdb==0.13.13

bandit==1.7.5
pip-audit==2.5.6
pipdeptree==2.9.0

coverage==7.2.7

# dependancies
asttokens==2.2.1
backcall==0.2.0
CacheControl==0.13.1
cfgv==3.3.1
cyclonedx-python-lib==2.7.1
decorator==5.1.1
distlib==0.3.6
django-stubs-ext==4.2.1
executing==1.2.0
filelock==3.12.0
gitdb==4.0.10
GitPython==3.1.31
html5lib==1.1
identify==2.5.24
ipython==8.14.0
jedi==0.18.2
markdown-it-py==2.2.0
matplotlib-inline==0.1.6
mdurl==0.1.2
msgpack==1.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
packageurl-python==0.11.1
packaging==23.1
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
pip-api==0.0.30
pip-requirements-parser==32.0.1
pip_audit==2.5.6
platformdirs==3.5.1
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
rich==13.4.1
ruff==0.0.272
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
stack-data==0.6.2
stevedore==5.1.0
toml==0.10.2
tomli==2.0.1
traitlets==5.9.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-urllib3==1.26.25.13
virtualenv==20.23.0
wcwidth==0.2.6
webencodings==0.5.1